requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
colorama==0.4.5
decorator==4.4.2
dj-database-url==1.0.0
Django==4.1.2
django-cors-headers==3.13.0
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
future==0.18.2
gunicorn==20.1.0
huepy==1.2.1
idna==3.4
imageio==2.22.1
imageio-ffmpeg==0.4.7
iniconfig==1.1.1
instabot==0.117.0
mock==4.0.3
moviepy==1.0.3
numpy==1.23.3
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
proglog==0.1.10
psycopg2==2.9.4
py==1.11.0
PyJWT==2.5.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.3
python-decouple==3.6
pytz==2022.4
requests==2.28.1
requests-toolbelt==0.9.1
responses==0.21.0
schedule==1.1.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.64.1
tzdata==2022.4
urllib3==1.26.12
whitenoise==6.2.0
